pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "pypdf2>=3.0.1",
    "reportlab>=4.4.1",
    "sqlalchemy>=2.0.41",
    "streamlit>=1.45.1",
]
